fix day3 exercise bug
The students are told to implement a Point class on this day. They are
given a Point class template to start from. That template has them
implement the __len__ magic method to return the length of the point as
if it were a vector. I.e. __len__ will return a float. Weirdly enough,
somewhere between __len__ returning and len() returning, python casts
the float to an int. I guess python really insists that len() returns an
int. All that's to say, we should use __len__ like we are. I just
renamed the __len__(self) method to length(self).

Pivotal Tracker:
https://www.pivotaltracker.com/projects/1433720/stories/113479209